cambios interface

+++ src/model/Amount.java
@@ -22,7 +22,7 @@ public class Amount {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
